+++ test/tools/router_check/test/route_tests.sh
@@ -3,8 +3,8 @@
 set -e
 
 # Router_check_tool binary path
-PATH_BIN="${TEST_RUNDIR}"/test/tools/router_check/router_check_tool
-
+PATH_BIN=/private/var/tmp/_bazel_jmahapatra/2fb333e5536650745e9f1244a12ec391/execroot/envoy/bazel-out/darwin-dbg/bin/test/tools/router_check/router_check_tool
+TEST_RUNDIR="/Users/jmahapatra/src/envoy"
 # Config json path
 PATH_CONFIG="${TEST_RUNDIR}"/test/tools/router_check/test/config
 
@@ -14,6 +14,17 @@ TESTS=("ContentType" "ClusterHeader" "HeaderMatchedRouting" "Redirect" "Redirect
 for t in "${TESTS[@]}"
 do
   TEST_OUTPUT=$("${PATH_BIN}" "${PATH_CONFIG}/${t}.yaml" "${PATH_CONFIG}/${t}.golden.json" "--details")
+done
+
+# Testing expected matches using --useproto
+# --useproto needs the test schema as a validation.proto message.
+for t in "${TESTS[@]}"
+do
+  sed -i '1s/^/{"tests":\n/' "${PATH_CONFIG}/${t}.golden.json"
+  sed -i "\$a}" "${PATH_CONFIG}/${t}.golden.json"
+  TEST_OUTPUT=$("${PATH_BIN}" "${PATH_CONFIG}/${t}.yaml" "${PATH_CONFIG}/${t}.golden.json" "--details" "--useproto")
+  sed -i '$ d' "${PATH_CONFIG}/${t}.golden.json"
+  sed -i '1d' "${PATH_CONFIG}/${t}.golden.json"
 done
 
 # Bad config file